Rename file extensions to hpp. Add new headers for C-style wrapper for kernel functions. Clean up of header guards and file layout. This commit covers MSD and dependencies. Branch does not yet compile.
