+++ include/device_MSD_plane_profile.hpp
@@ -1,5 +1,5 @@
-#ifndef ASTROACCELERATE_MSD_PLANE_PROFILE_H_
-#define ASTROACCELERATE_MSD_PLANE_PROFILE_H_
+#ifndef ASTRO_ACCELERATE_DEVICE_MSD_PLANE_PROFILE_HPP
+#define ASTRO_ACCELERATE_DEVICE_MSD_PLANE_PROFILE_HPP
 
 #include <vector>
