+++ include/device_corner_turn.hpp
@@ -1,5 +1,5 @@
-#ifndef ASTROACCELERATE_CORNERTURN_H_
-#define ASTROACCELERATE_CORNERTURN_H_
+#ifndef ASTRO_ACCELERATE_DEVICE_CORNER_TURN_HPP
+#define ASTRO_ACCELERATE_DEVICE_CORNER_TURN_HPP
 
 extern void corner_turn(unsigned short *d_input, float *d_output, int nchans, int nsamp);
 extern void corner_turn(float *d_input, float *d_output, int primary_size, int secondary_size);
