+++ include/device_single_FIR.hpp
@@ -1,5 +1,5 @@
-#ifndef ASTROACCELERATE_SINGLE_FIR_H_
-#define ASTROACCELERATE_SINGLE_FIR_H_
+#ifndef ASTRO_ACCELERATE_DEVICE_SINGLE_FIR_HPP
+#define ASTRO_ACCELERATE_DEVICE_SINGLE_FIR_HPP
 
 extern void PD_FIR_init(void);
 extern int PD_FIR(float *d_input, float *d_output, int nTaps, int nDMs, int nTimesamples);
