+++ include/helper_string.hpp
