--- include/host_rfi.h
@@ -1,7 +1,0 @@
-#ifndef ASTROACCELERATE_RFI_H_
-#define ASTROACCELERATE_RFI_H_
-
-void rfi(int nsamp, int nchans, unsigned short **input_buffer);
-
-#endif
-
