+++ include/host_rfi.hpp
@@ -1,0 +1,7 @@
+#ifndef ASTRO_ACCELERATE_HOST_RFI_HPP
+#define ASTRO_ACCELERATE_HOST_RFI_HPP
+
+void rfi(int nsamp, int nchans, unsigned short **input_buffer);
+
+#endif
+
