+++ src/device_bin.cu
@@ -1,84 +1,84 @@
-//#include <helper_cuda.h>
-// #include <omp.h>
-#include <time.h>
-#include <stdio.h>
-#include "headers/params.h"
-#include "device_binning_kernel.cu"
+#include "device_bin.hpp"
 
-//extern "C" void bin_gpu(float *bin_buffer, float *input_buffer, int nchans, int nsamp);
+void bin_gpu(unsigned short *d_input, float *d_output, int nchans, int nsamp) {
+  
+  int divisions_in_t = BINDIVINT;
+  int divisions_in_f = BINDIVINF;
+  int num_blocks_t = (int) ( ( nsamp + 1 ) / ( 2 * divisions_in_t ) );
+  int num_blocks_f = nchans / divisions_in_f;
+  
+  dim3 threads_per_block(divisions_in_t, divisions_in_f);
+  dim3 num_blocks(num_blocks_t, num_blocks_f);
+  //printf("\nDIVISOR:\t%f", (float)(nsamp)/(2*divisions_in_t));
+  //printf("\ndivisions_in_t:%d\tdivisions_in_f:%d",divisions_in_t, divisions_in_f);
+  //printf("\nnum_blocks_t:%d\tnum_blocks_f:%d\n",num_blocks_t,num_blocks_f);
+  
+  //cudaFuncSetCacheConfig(bin, cudaFuncCachePreferL1);
+  // clock_t start_t, end_t;
+  // start_t = clock();
+  //REPLACE WITH CALL TO C-WRAPPER FUNCTION
+  call_kernel_bin(num_blocks, threads_per_block, d_input, d_output, nsamp);
+	
+  //	getLastCudaError("Kernel execution failed");
 
-void bin_gpu(unsigned short *d_input, float *d_output, int nchans, int nsamp)
-{
+  int swap_divisions_in_t = CT;
+  int swap_divisions_in_f = CF;
+  int swap_num_blocks_t = nsamp / swap_divisions_in_t;
+  int swap_num_blocks_f = nchans / swap_divisions_in_f;
 
-	int divisions_in_t = BINDIVINT;
-	int divisions_in_f = BINDIVINF;
-	int num_blocks_t = (int) ( ( nsamp + 1 ) / ( 2 * divisions_in_t ) );
-	int num_blocks_f = nchans / divisions_in_f;
+  dim3 swap_threads_per_block(swap_divisions_in_t, swap_divisions_in_f);
+  dim3 swap_num_blocks(swap_num_blocks_t, swap_num_blocks_f);
 
-	dim3 threads_per_block(divisions_in_t, divisions_in_f);
-	dim3 num_blocks(num_blocks_t, num_blocks_f);
-	//printf("\nDIVISOR:\t%f", (float)(nsamp)/(2*divisions_in_t));
-	//printf("\ndivisions_in_t:%d\tdivisions_in_f:%d",divisions_in_t, divisions_in_f);
-	//printf("\nnum_blocks_t:%d\tnum_blocks_f:%d\n",num_blocks_t,num_blocks_f);
+  cudaDeviceSynchronize();
+  //REPLACE WITH CALL TO C-WRAPPER FUNCTION
+  call_kernel_swap(swap_num_blocks, swap_threads_per_block, d_input, d_output, nchans, nsamp);
+  cudaDeviceSynchronize();
 
-	//cudaFuncSetCacheConfig(bin, cudaFuncCachePreferL1);
-	// clock_t start_t, end_t;
-	// start_t = clock();
+  // end_t = clock();
+  // double time = double ( end_t - start_t ) / CLOCKS_PER_SEC;
+  //printf("\nPerformed Bin: %f (GPU estimate)", time);
+  //printf("\nGops based on %.2f ops per channel per tsamp: %f",14.0,((15.0*(divisions_in_t*divisions_in_f*num_blocks_t*num_blocks_f))/(time))/1000000000.0);
+  //printf("\nBN Device memory bandwidth in GB/s: %f", (2*(sizeof(float)+sizeof(unsigned short))*(divisions_in_t*divisions_in_f*num_blocks_t*num_blocks_f))/(time)/1000000000.0);
 
-	bin<<<num_blocks, threads_per_block>>>(d_input, d_output, nsamp);
-	//	getLastCudaError("Kernel execution failed");
-
-	int swap_divisions_in_t = CT;
-	int swap_divisions_in_f = CF;
-	int swap_num_blocks_t = nsamp / swap_divisions_in_t;
-	int swap_num_blocks_f = nchans / swap_divisions_in_f;
-
-	dim3 swap_threads_per_block(swap_divisions_in_t, swap_divisions_in_f);
-	dim3 swap_num_blocks(swap_num_blocks_t, swap_num_blocks_f);
-
-	cudaDeviceSynchronize();
-	swap<<<swap_num_blocks, swap_threads_per_block>>>(d_input, d_output, nchans, nsamp);
-	cudaDeviceSynchronize();
-
-	// end_t = clock();
-	// double time = double ( end_t - start_t ) / CLOCKS_PER_SEC;
-	//printf("\nPerformed Bin: %f (GPU estimate)", time);
-	//printf("\nGops based on %.2f ops per channel per tsamp: %f",14.0,((15.0*(divisions_in_t*divisions_in_f*num_blocks_t*num_blocks_f))/(time))/1000000000.0);
-	//printf("\nBN Device memory bandwidth in GB/s: %f", (2*(sizeof(float)+sizeof(unsigned short))*(divisions_in_t*divisions_in_f*num_blocks_t*num_blocks_f))/(time)/1000000000.0);
-
-	cudaMemset(d_output, 0, nchans * nsamp * sizeof(float));
-	//cudaMemcpy(input_buffer, bin_buffer, sizeof(float)*nchans*(nsamp/2), cudaMemcpyDeviceToDevice);
-	//	getLastCudaError("Kernel execution failed");
+  cudaMemset(d_output, 0, nchans * nsamp * sizeof(float));
+  //cudaMemcpy(input_buffer, bin_buffer, sizeof(float)*nchans*(nsamp/2), cudaMemcpyDeviceToDevice);
+  //	getLastCudaError("Kernel execution failed");
 }
 
 
-int GPU_DiT_v2_wrapper(float *d_input, float *d_output, int nDMs, int nTimesamples){
-	GpuTimer timer;
+int GPU_DiT_v2_wrapper(float *d_input, float *d_output, int nDMs, int nTimesamples) {
+  GpuTimer timer;
 	
-	//---------> Task specific
-	int ut;
+  //---------> Task specific
+  int ut;
 	
-	//---------> CUDA block and CUDA grid parameters
-	int nThreads = 2*WARP;
-	int nCUDAblocks_x=nTimesamples/(DIT_ELEMENTS_PER_THREAD*nThreads*2);
-	int nRest = nTimesamples%(DIT_ELEMENTS_PER_THREAD*nThreads*2);
-	if(nRest>0) nCUDAblocks_x++;
-	int nCUDAblocks_y=nDMs/DIT_YSTEP;
+  //---------> CUDA block and CUDA grid parameters
+  int nThreads = 2*WARP;
+  int nCUDAblocks_x=nTimesamples/(DIT_ELEMENTS_PER_THREAD*nThreads*2);
+  int nRest = nTimesamples%(DIT_ELEMENTS_PER_THREAD*nThreads*2);
+  if(nRest>0) nCUDAblocks_x++;
+  int nCUDAblocks_y=nDMs/DIT_YSTEP;
 	
-	dim3 gridSize(nCUDAblocks_x, nCUDAblocks_y, 1);
-	dim3 blockSize(nThreads, DIT_YSTEP, 1);
+  dim3 gridSize(nCUDAblocks_x, nCUDAblocks_y, 1);
+  dim3 blockSize(nThreads, DIT_YSTEP, 1);
 	
-	if( (nTimesamples>>1)==0 ) return(-1);
+  if( (nTimesamples>>1)==0 ) return(-1);
 	
-	// ----------------------------------------------->
-	// --------> Measured part (Pulse detection FIR)
-	//---------> Pulse detection FIR
-	DiT_GPU_v2<<<gridSize,blockSize>>>(d_input, d_output, nDMs, nTimesamples, (nTimesamples>>1));
-	// --------> Measured part (Pulse detection FIR)
-	// ----------------------------------------------->
+  // ----------------------------------------------->
+  // --------> Measured part (Pulse detection FIR)
+  //---------> Pulse detection FIR
+  call_kernel_DiT_GPU_v2(gridSize,
+			 blockSize,
+			 d_input,
+			 d_output,
+			 nDMs,
+			 nTimesamples,
+			 (nTimesamples>>1));
+  // --------> Measured part (Pulse detection FIR)
+  // ----------------------------------------------->
 	
-	ut=0;
-	return(ut);
+  ut=0;
+  return(ut);
 }
